feat(ffind): added quick-fix list support
Refactored the open_picker code to now take an actions table, which
includes all the operations that are available on a certain 'ffind'
picker, which is passed into 'M.open_picker' in place of the old
'on_select' parameter, which is now a mandatory entry in the table.
Defined two new optional entries 'to_qflist' and 'to_loclist', of which
'to_qflist' has actual implementations. Mapped 'to_qflist' to '<C-q>' and
'to_loclist' to '<C-l>'.